chore: bump 'tested up to' to 6.9
Also downgrade version for 'tested to' update.

To match the version currently live on WP.org (1.0.1):

https://wordpress.org/plugins/genesis-beta-tester/

This is the version that Circle automation is using for the remote
readme update:

https://github.com/studiopress/wp-svn-orb/blob/df9ee4b21b57fd6284edb598521f4c6ffa476bf7/src/scripts/env-vars-common.sh#L25

Our automation bumps the readme under tags/[version] on the remote.

But the version on WP.org is 1.0.1, not 1.0.2 (we only deployed 1.0.2 to
WPE servers):

https://wordpress.org/plugins/genesis-beta-tester/

So deployment to SVN succeeds but the remote version doesn't
update because we actually need to update tags/1.0.1, not tags/1.0.2,
which doesn't exist on the remote:

A    tags
Updated to revision 3410794.
Updating 'tags/1.0.2':
At revision 3410794.
There isn't a readme file at tags/1.0.2/readme.txt

https://app.circleci.com/pipelines/github/studiopress/genesis-beta-tester/70/workflows/03b051c3-9151-4d75-b83f-3ce59ec2ea05/jobs/93

So we have to temporarily downgrade the version to match the one on
WP.org and push that commit. Then automation will correctly update
the 'latest' tags/1.0.1 folder with the new readme.

After that, we can make a new commit to revert that version bump with
[skip ci] in the commit title or body so it does not redeploy.